DRYing of client settings for resources
