+++ pipedrive/resource_deals.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -56,30 +54,18 @@ func resourceDeals() *schema.Resource {
 }
 
 func resourceDealsCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	client := &http.Client{Timeout: 10 * time.Second}
-	apikey := m.(*Client).apitoken
-	baseurl := m.(*Client).baseurl
-	apiurl := fmt.Sprintf("%s/deals%s", baseurl, apikey)
-	payload := DealsBody(d)
-
 	// Warning or errors can be collected in a slice type
 	var diags diag.Diagnostics
+	path := "deals"
+	body := DealsBody(d)
+	resp, _, _, err := m.(*Client).SendRequest("POST", path, body, 201)
 
-	req, err := http.NewRequest("POST", apiurl, payload)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Accept", "application/json")
-
-	r, err := client.Do(req)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	defer r.Body.Close()
 
 	var result map[string]any
-	err = json.NewDecoder(r.Body).Decode(&result)
+	err = json.Unmarshal([]byte(resp), &result)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -95,31 +81,18 @@ func resourceDealsCreate(ctx context.Context, d *schema.ResourceData, m interfac
 }
 
 func resourceDealsRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	client := &http.Client{Timeout: 10 * time.Second}
-	id := d.Id()
-	apikey := m.(*Client).apitoken
-	baseurl := m.(*Client).baseurl
-	apiurl := fmt.Sprintf("%s/deals/%s%s", baseurl, id, apikey)
-
 	// Warning or errors can be collected in a slice type
 	var diags diag.Diagnostics
-	req, err := http.NewRequest("GET", apiurl, nil)
-	if err != nil {
-		return diag.FromErr(err)
-	}
+	id := d.Id()
+	path := "deals/" + id
+	resp, _, _, err := m.(*Client).SendRequest("GET", path, nil, 200)
 
-	r, err := client.Do(req)
-	if r.StatusCode == 404 && err != nil {
-		d.SetId("")
-		diag.FromErr(err)
-	}
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	defer r.Body.Close()
 
 	var result map[string]any
-	err = json.NewDecoder(r.Body).Decode(&result)
+	err = json.Unmarshal([]byte(resp), &result)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -136,54 +109,31 @@ func resourceDealsRead(ctx context.Context, d *schema.ResourceData, m interface{
 }
 
 func resourceDealsUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	client := &http.Client{Timeout: 10 * time.Second}
 	id := d.Id()
-	apikey := m.(*Client).apitoken
-	baseurl := m.(*Client).baseurl
-	apiurl := fmt.Sprintf("%s/deals/%s%s", baseurl, id, apikey)
-	payload := DealsBody(d)
-
-	req, err := http.NewRequest("PUT", apiurl, payload)
+	path := "deals/" + id
+	body := DealsBody(d)
+	_, _, _, err := m.(*Client).SendRequest("PUT", path, body, 200)
 
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Accept", "application/json")
-
-	r, err := client.Do(req)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	defer r.Body.Close()
 
 	return resourceDealsRead(ctx, d, m)
-
 }
 
 func resourceDealsDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	client := &http.Client{Timeout: 10 * time.Second}
-	id := d.Id()
-	apikey := m.(*Client).apitoken
-	baseurl := m.(*Client).baseurl
-	apiurl := fmt.Sprintf("%s/deals/%s%s", baseurl, id, apikey)
-
 	// Warning or errors can be collected in a slice type
 	var diags diag.Diagnostics
-	req, err := http.NewRequest("DELETE", apiurl, nil)
+	id := d.Id()
+	path := "deals/" + id
+	resp, _, _, err := m.(*Client).SendRequest("DELETE", path, nil, 200)
 
 	if err != nil {
 		return diag.FromErr(err)
 	}
-
-	r, err := client.Do(req)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	defer r.Body.Close()
 
 	var result map[string]any
-	err = json.NewDecoder(r.Body).Decode(&result)
+	err = json.Unmarshal([]byte(resp), &result)
 	if err != nil {
 		return diag.FromErr(err)
 	}
